Update v1.1.0
Added a DDoS page, which you can also customize and/or
manually/forcefully trigger. There really is no difference between it
and the downtime page, but if you want to show a specific page in the
event you get DDoS'ed and it goes through Cloudflare then you can direct
all the requests to this serverless script instead of your servers.